Back_end/requirements.txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
openai
